content/base/test/test_bug333198.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=333198
-->
<head>
  <title>Test for Bug 333198</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<iframe id="ifr"></iframe><br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=333198">Mozilla Bug 333198</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 333198 **/

var eventCount = 0;
function clickHandler() {
  ++eventCount;
}

function suppressEvents(suppress) {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
        .getInterface(Components.interfaces.nsIDOMWindowUtils)
        .suppressEventHandling(suppress);
}

function sendEvents() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  windowUtils = window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                      .getInterface(Components.interfaces.nsIDOMWindowUtils);
  windowUtils.sendMouseEvent("mousedown", 1, 1, 0, 1, 0);
  windowUtils.sendMouseEvent("mouseup", 1, 1, 0, 1, 0);

  iframeUtils = document.getElementById("ifr").contentWindow
                        .QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                        .getInterface(Components.interfaces.nsIDOMWindowUtils);
  iframeUtils.sendMouseEvent("mousedown", 1, 1, 0, 1, 0);
  iframeUtils.sendMouseEvent("mouseup", 1, 1, 0, 1, 0);
}

function runTest() {
  window.addEventListener("click", clickHandler, true);
  var ifr = document.getElementById("ifr")
  ifr.contentWindow.addEventListener("click", clickHandler, true);
  sendEvents();
  is(eventCount, 2, "Wrong event count(1)");
  suppressEvents(true);
  sendEvents();
  is(eventCount, 2, "Wrong event count(2)");
  suppressEvents(false);
  sendEvents();
  is(eventCount, 4, "Wrong event count(2)");
  if (eventCount != 4)
  alert(eventCount);
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
